+++ packages/prefabs/src/index.ts
@@ -16,3 +16,4 @@ export * from './area.js';
 export * from './systems.js';
 export * from './commands.js';
 export * from './queries.js';
+export * from './suggest.js';